Corrections, investigation d'un crash.

+++ logiciels/kuri/adn/genere_lexemes.cc
@@ -326,9 +326,9 @@ struct Lexeme {
     dls::vue_chaine_compacte chaine{};
 
 	union {
-		unsigned long long valeur_entiere;
+        uint64_t valeur_entiere;
 		double valeur_reelle;
-		long index_chaine;
+        int64_t index_chaine;
 		IdentifiantCode *ident;
 	};
